requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
transformers>=4.35.2
torch>=2.6.0
scikit-learn>=1.3.2
numpy>=1.24.3
pandas>=2.1.3
python-multipart>=0.0.6
sentence-transformers>=2.2.2
spacy>=3.7.2
nltk>=3.8.1
python-docx>=1.1.0
PyPDF2>=3.0.1 